Add flag to wait for migrations to complete (or not)
This makes the init container that waits for migrations optional via a
`waitForMigrations` flag. If disabled, it does not render the
`initContainer` alongside the role and role binding it uses.

In tools such as ArgoCD, the waves (which can be set throught he helm
annotations we already have) ensure that a resource is applied before
another one. So this waiting init container may not be needed. Let's
make it configurable for deployments that don't require this.

Note that the option defaults to `true` as to keep the current logic and
functionality as it is.