fix: set a different component label on migration Job
If the migration Job has the same set of label as the ones referenced in
the Deployment match labels then the HorizontalPodAutoscaler will match
the Job containers as being part of the replicaset.

This addresses this issue by setting the component lable to "migration"
on the migration Job and the migration Job Pods.

The pattern for setting the component label is now better aligned with
other open-source Helm charts.

Signed-off-by: Alexandre-P <alexandre.picosson@owkin.com>